refactor: spark 관련 코드 삭제
