--- src/main/java/chess/web/dao/DbConnector.java
@@ -1,22 +1,0 @@
-package chess.web.dao;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-public class DbConnector {
-
-    private static final String URL = "jdbc:mysql://localhost:3306/chess";
-    private static final String USER = "user";
-    private static final String PASSWORD = "password";
-
-    public static Connection getConnection() {
-        Connection connection;
-        try {
-            connection = DriverManager.getConnection(URL, USER, PASSWORD);
-        } catch (SQLException e) {
-            throw new IllegalArgumentException("DB에 연결할 수 없습니다.");
-        }
-        return connection;
-    }
-}
